+++ lib/views/settings/laboratory/game/mini_game/components/orc.dart
@@ -10,7 +10,8 @@ import 'boss.dart';
 
 double tileSize = 20.0;
 
-class Orc extends SimpleEnemy with ObjectCollision, AutomaticRandomMovement {
+class Orc extends SimpleEnemy
+    with ObjectCollision, AutomaticRandomMovement, UseBarLife {
   bool canMove = true;
 
   Orc(Vector2 position)
@@ -48,6 +49,24 @@ class Orc extends SimpleEnemy with ObjectCollision, AutomaticRandomMovement {
         ],
       ),
     );
+
+    /// 生命条
+    setupBarLife(
+      size: Vector2(tileSize * 1.5, tileSize / 5),
+      barLifePosition: BarLifePorition.top,
+      showLifeText: false,
+      margin: 0,
+      borderWidth: 2,
+      borderColor: Colors.white.withOpacity(0.5),
+      borderRadius: BorderRadius.circular(2),
+      offset: Vector2(0, tileSize * 0.5),
+    );
+  }
+
+  /// 渲染
+  @override
+  void render(Canvas canvas) {
+    super.render(canvas);
   }
 
   /// 碰撞触发
@@ -61,25 +80,6 @@ class Orc extends SimpleEnemy with ObjectCollision, AutomaticRandomMovement {
       active = false;
     }
     return active;
-  }
-
-  @override
-  void render(Canvas canvas) {
-    if (!isDead) {
-      /// 生命条
-      drawDefaultLifeBar(
-        canvas,
-        margin: 0,
-        width: tileSize * 1.5,
-        height: tileSize / 5,
-        borderRadius: BorderRadius.circular(2),
-        align: Offset(
-          tileSize * 1.7,
-          tileSize / 4,
-        ),
-      );
-    }
-    super.render(canvas);
   }
 
   @override
@@ -244,7 +244,7 @@ class Orc extends SimpleEnemy with ObjectCollision, AutomaticRandomMovement {
   void _execAttack() {
     simpleAttackMelee(
       damage: 10,
-      size: Vector2.all(tileSize * 1.5),
+      size: Vector2.all(tileSize * 5),
       interval: 800,
       execute: () {
         _addAttackAnimation();